Re-rendering issue (#1330)
* removed unused import

* formatting files

* re rending of table of contents

* file formatting

* removed a console.log

* authenticating routes
