+++ src/components/Wiki/WikiPage/WikiTableOfContents.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect } from 'react'
+import React, { useEffect, useRef, useState } from 'react'
 import {
   VStack,
   Text,
@@ -36,13 +36,13 @@ const WikiTableOfContents = ({ isAlertAtTop }: WikiTableOfContentsProps) => {
   const { colorMode } = useColorMode()
   const { isOpen, onToggle } = useDisclosure()
   const isDefaultOpen = useBreakpointValue({ base: false, xs: true, xl: false })
-  const [activeId, setActiveId] = React.useState<string | null>()
+  const [activeId, setActiveId] = useState<string | null>()
 
   // the below ref is used to store all the heading that are in view
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  const headingElementsRef: any = React.useRef({})
+  const headingElementsRef: any = useRef({})
 
-  React.useEffect(() => {
+  useEffect(() => {
     // get all the heading elements
     const headingElements = Array.from(
       document.querySelectorAll('h1, h2, h3, h4, h5, h6'),
@@ -97,88 +97,89 @@ const WikiTableOfContents = ({ isAlertAtTop }: WikiTableOfContentsProps) => {
     return () => observer.disconnect()
   }, [setActiveId, toc])
 
-  React.useEffect(() => {
+  useEffect(() => {
     if (!activeId) setActiveId(toc[0]?.id)
   }, [activeId, toc])
 
-  if (isOpen === isDefaultOpen) {
-    return (
-      <VStack
-        display={{ base: 'none', xl: 'block' }}
-        borderLeftWidth="1px"
-        w="20vw"
-        px={6}
-        py="30px"
-        ml="0 !important"
-      >
-        <VStack
-          spacing={4}
-          align="start"
-          position="sticky"
-          top="calc(70px + 30px + 2px)"
-        >
-          <Flex w="100%" justify="end">
-            <IconButton
-              aria-label="Toggle Table of Contents"
-              icon={<RiMenu3Fill />}
-              onClick={onToggle}
-            />
-          </Flex>
-          <StaticContent>
-            <VStack
-              as="nav"
-              spacing={4}
-              h="calc(100vh - (70px + 90px))"
-              overflowY="scroll"
-              pr={4}
-              w="100%"
-              align="start"
-              css={{
-                '&::-webkit-scrollbar': {
-                  width: '4px',
-                },
-                '&::-webkit-scrollbar-track': {
-                  width: '6px',
-                },
-                '&::-webkit-scrollbar-thumb': {
-                  background: colorMode === 'light' ? '#0000002a' : '#3f444e',
-                  borderRadius: '24px',
-                },
-              }}
-            >
-              {toc.map(({ level, id, title }) => (
-                <Box key={id} pl={`calc(${(level - 1) * 20}px)`}>
-                  <Text
-                    color={activeId === id ? 'brandLinkColor' : 'unset'}
-                    boxShadow={
-                      activeId === id ? '-2px 0px 0px 0px #ff5caa' : '0'
-                    }
-                    outlineColor="brandLinkColor"
-                    pl={2}
-                  >
-                    <Link href={`#${id}`}>{title}</Link>
-                  </Text>
-                </Box>
-              ))}
-            </VStack>
-          </StaticContent>
-        </VStack>
-      </VStack>
-    )
-  }
   return (
-    <Box
-      display={{ base: 'none', md: 'block' }}
-      pos="absolute"
-      right="24px"
-      top={`calc(70px + ${isAlertAtTop ? '80px' : '32px'})`}
-    >
-      <IconButton
-        aria-label="Toggle Table of Contents"
-        icon={<RiMenu3Fill />}
-        onClick={onToggle}
-      />
-    </Box>
+    <>
+      {isOpen === isDefaultOpen ? (
+        <VStack
+          display={{ base: 'none', xl: 'block' }}
+          borderLeftWidth="1px"
+          w="20vw"
+          px={6}
+          py="30px"
+          ml="0 !important"
+        >
+          <VStack
+            spacing={4}
+            align="start"
+            position="sticky"
+            top="calc(70px + 30px + 2px)"
+          >
+            <Flex w="100%" justify="end">
+              <IconButton
+                aria-label="Toggle Table of Contents"
+                icon={<RiMenu3Fill />}
+                onClick={onToggle}
+              />
+            </Flex>
+            <StaticContent>
+              <VStack
+                as="nav"
+                spacing={4}
+                h="calc(100vh - (70px + 90px))"
+                overflowY="scroll"
+                pr={4}
+                w="100%"
+                align="start"
+                css={{
+                  '&::-webkit-scrollbar': {
+                    width: '4px',
+                  },
+                  '&::-webkit-scrollbar-track': {
+                    width: '6px',
+                  },
+                  '&::-webkit-scrollbar-thumb': {
+                    background: colorMode === 'light' ? '#0000002a' : '#3f444e',
+                    borderRadius: '24px',
+                  },
+                }}
+              >
+                {toc.map(({ level, id, title }) => (
+                  <Box key={id} pl={`calc(${(level - 1) * 20}px)`}>
+                    <Text
+                      color={activeId === id ? 'brandLinkColor' : 'unset'}
+                      boxShadow={
+                        activeId === id ? '-2px 0px 0px 0px #ff5caa' : '0'
+                      }
+                      outlineColor="brandLinkColor"
+                      pl={2}
+                    >
+                      <Link href={`#${id}`}>{title}</Link>
+                    </Text>
+                  </Box>
+                ))}
+              </VStack>
+            </StaticContent>
+          </VStack>
+        </VStack>
+      ) : (
+        <Box
+          display={{ base: 'none', md: 'block' }}
+          pos="absolute"
+          right="24px"
+          top={`calc(70px + ${isAlertAtTop ? '80px' : '32px'})`}
+        >
+          <IconButton
+            aria-label="Toggle Table of Contents"
+            icon={<RiMenu3Fill />}
+            onClick={onToggle}
+          />
+        </Box>
+      )}
+    </>
   )
 }
