+++ src/components/WrapperRoutes/AuthenticatedRoute.tsx
@@ -1,7 +1,6 @@
 import { useRouter } from 'next/router'
-import React, { useContext, useEffect } from 'react'
+import React, { useEffect } from 'react'
 import { useAccount } from 'wagmi'
-import { WagmiStatusContext } from '../Wagmi/DynamicWagmiProvider'
 
 const WagmiLoadedAuthenticatedRoute = ({
   WrappedComponent,
@@ -31,20 +30,12 @@ export const authenticatedRoute = <P extends object>(
   WrappedComponent: () => JSX.Element | null,
 ) => {
   const AuthenticatedRoute = (props: P) => {
-    const { isWagmiWrapped, setIsWagmiWrapped } = useContext(WagmiStatusContext)
-
-    useEffect(() => {
-      setIsWagmiWrapped(true)
-    }, [setIsWagmiWrapped])
-
     return (
       <>
-        {isWagmiWrapped && (
-          <WagmiLoadedAuthenticatedRoute
-            WrappedComponent={WrappedComponent}
-            {...props}
-          />
-        )}
+        <WagmiLoadedAuthenticatedRoute
+          WrappedComponent={WrappedComponent}
+          {...props}
+        />
       </>
     )
   }
