+++ src/pages/_app.tsx
@@ -11,15 +11,23 @@ import { store } from '@/store/store'
 import NextNProgress from 'nextjs-progressbar'
 import { pageView } from '@/utils/googleAnalytics'
 import '../utils/i18n'
-import { DynamicWagmiProvider } from '@/components/Wagmi/DynamicWagmiProvider'
+import { WagmiConfig, createConfig } from 'wagmi'
 import { Montserrat } from '@next/font/google'
 import chakraTheme from '../theme'
+import { connectors, publicClient, webSocketPublicClient } from '@/config/wagmi'
 
 const { ToastContainer } = createStandaloneToast()
 
 type EpAppProps = Omit<AppProps, 'Component'> & {
   Component: AppProps['Component'] & { noFooter?: boolean }
 }
+
+const client = createConfig({
+  autoConnect: true,
+  connectors,
+  publicClient,
+  webSocketPublicClient,
+})
 
 export const montserrat = Montserrat({
   subsets: ['latin'],
@@ -45,11 +53,11 @@ const App = ({ Component, pageProps, router }: EpAppProps) => {
       <SEOHeader router={router} />
       <ReduxProvider store={store}>
         <ChakraProvider resetCSS theme={chakraTheme}>
-          <DynamicWagmiProvider>
+          <WagmiConfig config={client}>
             <Layout noFooter={Component.noFooter}>
               <Component {...pageProps} />
             </Layout>
-          </DynamicWagmiProvider>
+          </WagmiConfig>
         </ChakraProvider>
       </ReduxProvider>
       <ToastContainer />
